requirements.txt
Django===3.2
djangorestframework===3.15.0
pyshorteners===1.0.1